+++ ethereum/statetransition/src/test/java/tech/pegasys/teku/statetransition/attestation/AggregatingAttestationPoolTest.java
@@ -29,12 +29,9 @@ import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
 import org.mockito.ArgumentMatchers;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.Spec;
-import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.TestSpecFactory;
 import tech.pegasys.teku.spec.datastructures.attestation.ValidateableAttestation;
 import tech.pegasys.teku.spec.datastructures.operations.Attestation;
@@ -117,7 +114,7 @@ class AggregatingAttestationPoolTest {
   public void getAttestationsForBlock_shouldReturnEmptyListWhenNoAttestationsAvailable() {
     when(mockSpec.validateAttestation(any(), any())).thenReturn(Optional.empty());
 
-    final var stateAtBlockSlot = dataStructureUtil.randomBeaconState();
+    final BeaconState stateAtBlockSlot = dataStructureUtil.randomBeaconState();
 
     assertThat(aggregatingPool.getAttestationsForBlock(stateAtBlockSlot, forkChecker)).isEmpty();
   }
@@ -131,7 +128,7 @@ class AggregatingAttestationPoolTest {
     when(mockSpec.validateAttestation(any(), any()))
         .thenReturn(Optional.of(AttestationInvalidReason.SLOT_NOT_IN_EPOCH));
 
-    final var stateAtBlockSlot = dataStructureUtil.randomBeaconState();
+    final BeaconState stateAtBlockSlot = dataStructureUtil.randomBeaconState();
 
     assertThat(aggregatingPool.getAttestationsForBlock(stateAtBlockSlot, forkChecker)).isEmpty();
   }
@@ -179,7 +176,7 @@ class AggregatingAttestationPoolTest {
     final Attestation attestation1 = addAttestationFromValidators(attestationData, 1, 2);
     final Attestation attestation2 = addAttestationFromValidators(attestationData, 3, 4);
 
-    final var stateAtBlockSlot = dataStructureUtil.randomBeaconState();
+    final BeaconState stateAtBlockSlot = dataStructureUtil.randomBeaconState();
 
     assertThat(aggregatingPool.getAttestationsForBlock(stateAtBlockSlot, forkChecker))
         .containsExactly(aggregateAttestations(attestation1, attestation2));
@@ -193,7 +190,7 @@ class AggregatingAttestationPoolTest {
     final Attestation attestation3 =
         addAttestationFromValidators(dataStructureUtil.randomAttestationData(), 3, 4);
 
-    final var stateAtBlockSlot = dataStructureUtil.randomBeaconState();
+    final BeaconState stateAtBlockSlot = dataStructureUtil.randomBeaconState();
 
     assertThat(aggregatingPool.getAttestationsForBlock(stateAtBlockSlot, forkChecker))
         .containsExactlyInAnyOrder(aggregateAttestations(attestation1, attestation2), attestation3);
@@ -282,7 +279,7 @@ class AggregatingAttestationPoolTest {
     final Attestation preserveAttestation =
         addAttestationFromValidators(preserveAttestationData, 2);
 
-    final var stateAtBlockSlot = dataStructureUtil.randomBeaconState();
+    final BeaconState stateAtBlockSlot = dataStructureUtil.randomBeaconState();
 
     assertThat(aggregatingPool.getSize()).isEqualTo(2);
     aggregatingPool.onSlot(
@@ -485,44 +482,16 @@ class AggregatingAttestationPoolTest {
     return new Attestation(bitlist, data, dataStructureUtil.randomSignature());
   }
 
-  @ParameterizedTest
-  @EnumSource(
-      value = SpecMilestone.class,
-      names = {"MERGE"},
-      mode = EnumSource.Mode.EXCLUDE)
-  void getAttestationsForBlock_inAltairShouldNotIncludeWorthlessAttestations(
-      SpecMilestone milestone) {
-    switch (milestone) {
-      case PHASE0:
-        spec = TestSpecFactory.createMainnetPhase0();
-        break;
-      case ALTAIR:
-        spec = TestSpecFactory.createMainnetAltair();
-        break;
-      default:
-        throw new IllegalStateException("unsupported milestone");
-    }
-
+  @Test
+  void getAttestationsForBlock_inAltairShouldNotIncludeWorthlessAttestations() {
+    spec = TestSpecFactory.createMainnetAltair();
     dataStructureUtil = new DataStructureUtil(spec);
 
-    final var referenceAttestationData = dataStructureUtil.randomAttestationData(UInt64.valueOf(1));
+    final AttestationData referenceAttestationData =
+        dataStructureUtil.randomAttestationData(UInt64.valueOf(1));
     final BeaconState stateAtBlockSlot = dataStructureUtil.randomBeaconState(UInt64.valueOf(10));
 
-    final var correctTarget =
-        new Checkpoint(
-            dataStructureUtil.randomEpoch(),
-            spec.atSlot(stateAtBlockSlot.getSlot())
-                .beaconStateAccessors()
-                .getBlockRootAtSlot(stateAtBlockSlot, ZERO));
-    final var wrongTarget = dataStructureUtil.randomCheckpoint();
-
-    final AttestationData attestationData1_SlotOld_TargetOK =
-        new AttestationData(
-            UInt64.valueOf(1),
-            referenceAttestationData.getIndex().plus(1),
-            referenceAttestationData.getBeacon_block_root(),
-            referenceAttestationData.getSource(),
-            correctTarget);
+    final Checkpoint wrongTarget = dataStructureUtil.randomCheckpoint();
 
     final AttestationData attestationData2_SlotOld_TargetWrong =
         new AttestationData(
@@ -532,42 +501,8 @@ class AggregatingAttestationPoolTest {
             referenceAttestationData.getSource(),
             wrongTarget);
 
-    final AttestationData attestationData3_SlotOK_TargetWrong =
-        new AttestationData(
-            UInt64.valueOf(5),
-            referenceAttestationData.getIndex().plus(3),
-            referenceAttestationData.getBeacon_block_root(),
-            referenceAttestationData.getSource(),
-            wrongTarget);
+    addAttestationFromValidators(attestationData2_SlotOld_TargetWrong, 1, 2);
 
-    final AttestationData attestationData4_SlotOK_TargetOK =
-        new AttestationData(
-            UInt64.valueOf(7),
-            referenceAttestationData.getIndex().plus(4),
-            referenceAttestationData.getBeacon_block_root(),
-            referenceAttestationData.getSource(),
-            correctTarget);
-
-    final Attestation attestation1 =
-        addAttestationFromValidators(attestationData1_SlotOld_TargetOK, 1, 2);
-    final Attestation attestation2 =
-        addAttestationFromValidators(attestationData2_SlotOld_TargetWrong, 3, 4);
-    final Attestation attestation3 =
-        addAttestationFromValidators(attestationData3_SlotOK_TargetWrong, 5, 6);
-    final Attestation attestation4 =
-        addAttestationFromValidators(attestationData4_SlotOK_TargetOK, 7, 8);
-
-    Attestation[] expectedAttestations;
-
-    if (milestone.equals(SpecMilestone.PHASE0)) {
-      expectedAttestations =
-          new Attestation[] {attestation4, attestation3, attestation2, attestation1};
-    } else {
-      // we expect attestation2 to be discarded
-      expectedAttestations = new Attestation[] {attestation4, attestation3, attestation1};
-    }
-
-    assertThat(aggregatingPool.getAttestationsForBlock(stateAtBlockSlot, forkChecker))
-        .containsExactly(expectedAttestations);
+    assertThat(aggregatingPool.getAttestationsForBlock(stateAtBlockSlot, forkChecker)).isEmpty();
   }
 }
